+++ client/my-sites/email/gsuite-purchase-cta/test/index.js
@@ -15,7 +15,7 @@ import { GSuitePurchaseCta } from '../';
 // components to mock
 jest.mock( 'components/email-verification/email-verification-gate', () => 'EmailVerificationGate' );
 jest.mock( 'my-sites/email/gsuite-purchase-cta/sku-info', () => 'GSuitePurchaseCtaSkuInfo' );
-jest.mock( 'my-sites/email/gsuite-purchase-features', () => 'GSuitePurchaseFeatures' );
+jest.mock( 'my-sites/email/gsuite-features', () => 'GSuiteFeatures' );
 jest.mock( 'components/data/query-products-list', () => 'QueryProductsList' );
 
 describe( 'GSuitePurchaseCta', () => {
